requirements.txt
Django==3.0.2
djangorestframework==3.12.1
Pillow==8.0.1
